Fix linker errors, add copyright infromation.
