+++ integration_test/CMakeLists.txt
@@ -5,5 +5,5 @@ project(asepriteConvertTest)
 include_directories(${AsepriteRgbaConvert_SOURCE_DIR})
 include_directories(${ZLIB_INCLUDE_DIRS})
 add_executable(asepriteConvertTest main.cpp)
-add_dependencies(asepriteConvertTest AsepriteRgbaConvert)
+target_link_libraries(asepriteConvertTest AsepriteRgbaConvert)
 target_link_libraries(asepriteConvertTest ${ZLIB_LIBRARIES})
